# openapi.json
{
  "components": {
    "schemas": {
...
      "EbookDocumentRead": {
        "description": "HTTP/tool representation of a document Tether has seen progress for.\n\n`finished` collapses the internal `finished_captured_at` stamp to the only\nfact a surface needs \u2014 whether the one finished Memory has been minted.\n\n>>> EbookDocumentRead(document_hash=\"abc\", title=None, finished=False).finished\nFalse",
        "properties": {
          "document_hash": {
            "title": "Document Hash",
            "type": "string"
          },
          "finished": {
            "title": "Finished",
            "type": "boolean"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "document_hash",
          "title",
          "finished"
        ],
        "title": "EbookDocumentRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LabelEbookRequest": {
        "description": "Body for attaching a human title to a document hash.",
        "properties": {
          "document_hash": {
            "minLength": 1,
            "title": "Document Hash",
            "type": "string"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "document_hash",
          "title"
        ],
        "title": "LabelEbookRequest",
        "type": "object"
      },
...
      "MatchEbookFilenameRequest": {
        "description": "Body for labeling the document a filename hashes to.",
        "properties": {
          "filename": {
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          }
        },
        "required": [
          "filename"
        ],
        "title": "MatchEbookFilenameRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Tether",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/ebooks/label": {
      "post": {
        "description": "Attach a human title to a document hash.",
        "operationId": "label_ebook_api_ebooks_label_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelEbookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EbookDocumentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Label Ebook"
      }
    },
    "/api/ebooks/match-filename": {
      "post": {
        "description": "Label the document a filename hashes to, titled from the filename.",
        "operationId": "match_ebook_filename_api_ebooks_match_filename_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchEbookFilenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EbookDocumentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Match Ebook Filename"
      }
    },
...
  }
}